+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# [](https://github.com/dragon-ball-p/dragon-ball/compare/v0.0.4-alpha.0...v) (2022-06-22)
+# [](https://github.com/dragon-ball-p/dragon-ball/compare/v0.0.4-alpha.0...v) (2022-06-23)
 
 
 ### Bug Fixes
@@ -8,6 +8,7 @@
 
 ### Features
 
+* 支持 link 按钮 ([b8ce387](https://github.com/dragon-ball-p/dragon-ball/commit/b8ce3875d58cdde59cb503a30ed45940f97d8d11))
 * form ([7c7a921](https://github.com/dragon-ball-p/dragon-ball/commit/7c7a9218d6fe8bf9a75bfee95fbfca3853f02504))
 * form WIP ([46d5cc4](https://github.com/dragon-ball-p/dragon-ball/commit/46d5cc488f3d88e1c74be62c6a9c36d07e3161c3))
 * wIP form, 支持 input checkbox, radio ([0be8ea2](https://github.com/dragon-ball-p/dragon-ball/commit/0be8ea2dff84c6c81e7104c962c330c292b50275))