+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# [](https://github.com/dragon-ball-p/dragon-ball/compare/v0.0.4-alpha.0...v) (2022-06-24)
+# [](https://github.com/dragon-ball-p/dragon-ball/compare/v0.0.4-alpha.0...v) (2022-08-12)
 
 
 ### Bug Fixes
@@ -12,6 +12,11 @@
 * 支持 link 按钮 ([b8ce387](https://github.com/dragon-ball-p/dragon-ball/commit/b8ce3875d58cdde59cb503a30ed45940f97d8d11))
 * form ([7c7a921](https://github.com/dragon-ball-p/dragon-ball/commit/7c7a9218d6fe8bf9a75bfee95fbfca3853f02504))
 * form WIP ([46d5cc4](https://github.com/dragon-ball-p/dragon-ball/commit/46d5cc488f3d88e1c74be62c6a9c36d07e3161c3))
+* **form:** 初步实现添加 rule，并触发重渲染 Item ([5985374](https://github.com/dragon-ball-p/dragon-ball/commit/5985374cf2fe171bae15dc013962147d591ab835))
+* **form:** 添加 useForm，并支持类组件使用 form ([ba43560](https://github.com/dragon-ball-p/dragon-ball/commit/ba43560e64bf04d1fcb3525fa4e9ab4ac465c886))
+* **form:** 添加初始值，错误显示，清除销毁的Item ([064003d](https://github.com/dragon-ball-p/dragon-ball/commit/064003d6035c9cfefcaaba09254f70c547e07b45))
+* **form:** 重写 form ([3415dd2](https://github.com/dragon-ball-p/dragon-ball/commit/3415dd2ff1ea007a70dee3ff3e3dec02f535fe69))
+* **input:** export input ([896e24c](https://github.com/dragon-ball-p/dragon-ball/commit/896e24c5e76104c3bd8dafe7b4d152748a0db94c))
 * menu 菜单功能完善 ([dbeaf26](https://github.com/dragon-ball-p/dragon-ball/commit/dbeaf26a8fc74d448418367635dd8fbeea251ce5))
 * **menu:** 为 MenuItem 自动添加 index ([19629e4](https://github.com/dragon-ball-p/dragon-ball/commit/19629e496fc5c8c207d4f1ef92af11b708113028))
 * **menu:** wIP: 初步实现 Sub Menu ([07b4d21](https://github.com/dragon-ball-p/dragon-ball/commit/07b4d21a18add2d1929915df9c41b27dd117d52e))