+++ ChangeLog.md
@@ -1,14 +1,17 @@
-# [](https://github.com/dragon-ball-p/dragon-ball/compare/v0.0.4-alpha.0...v) (2022-08-12)
+# [](https://github.com/dragon-ball-p/dragon-ball/compare/v0.0.4-alpha.0...v) (2022-08-13)
 
 
 ### Bug Fixes
 
+* 修复语法错误 ([891d11a](https://github.com/dragon-ball-p/dragon-ball/commit/891d11a081b1ab57964a7c3c9cd72421c7678d6f))
 * fix Notification ([f639f6b](https://github.com/dragon-ball-p/dragon-ball/commit/f639f6bee86b54069d8df36761cb308906a3cad7))
 
 
 ### Features
 
 * 初步实现 Menu ([6865490](https://github.com/dragon-ball-p/dragon-ball/commit/68654907fe082ceb2409dcb805ac1a79e758e43a))
+* 合并组件库 ([cd8ed44](https://github.com/dragon-ball-p/dragon-ball/commit/cd8ed44d99b9345128bbaf37f7a21052fa110fb7))
+* 提取 CSS ([dfe3604](https://github.com/dragon-ball-p/dragon-ball/commit/dfe3604d8633730b0ff4dca625c4bf1963de1375))
 * 支持 link 按钮 ([b8ce387](https://github.com/dragon-ball-p/dragon-ball/commit/b8ce3875d58cdde59cb503a30ed45940f97d8d11))
 * form ([7c7a921](https://github.com/dragon-ball-p/dragon-ball/commit/7c7a9218d6fe8bf9a75bfee95fbfca3853f02504))
 * form WIP ([46d5cc4](https://github.com/dragon-ball-p/dragon-ball/commit/46d5cc488f3d88e1c74be62c6a9c36d07e3161c3))
@@ -20,6 +23,7 @@
 * menu 菜单功能完善 ([dbeaf26](https://github.com/dragon-ball-p/dragon-ball/commit/dbeaf26a8fc74d448418367635dd8fbeea251ce5))
 * **menu:** 为 MenuItem 自动添加 index ([19629e4](https://github.com/dragon-ball-p/dragon-ball/commit/19629e496fc5c8c207d4f1ef92af11b708113028))
 * **menu:** wIP: 初步实现 Sub Menu ([07b4d21](https://github.com/dragon-ball-p/dragon-ball/commit/07b4d21a18add2d1929915df9c41b27dd117d52e))
+* popover 的样式 ([5c10d60](https://github.com/dragon-ball-p/dragon-ball/commit/5c10d6069fdaacc7f9de1704f694e6411d89b8be))
 * wIP form, 支持 input checkbox, radio ([0be8ea2](https://github.com/dragon-ball-p/dragon-ball/commit/0be8ea2dff84c6c81e7104c962c330c292b50275))